// src/app/models/master/masOccupationData.ts
/**
 * Generated by orval v6.23.0 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */
import type { MasOccupationSkillData } from './masOccupationSkillData';

export interface MasOccupationData {
    /** สถานะการใช้งาน */
    activeFlag?: boolean;
    first?: number;
    mode?: string;
    /** รหัสอาชีพ */
    occupationCode: string;
    /** กลุ่มอาชีพ */
    occupationGroupId?: number;
    /** custom ชื่อกลุ่มอาชีพ (ภาษาอังกฤษ) */
    occupationGroupNameEn?: string;
    /** custom ชื่อกลุ่มอาชีพ (ภาษาไทย) */
    occupationGroupNameTh?: string;
    /** ข้อมูลอาชีพ */
    occupationId?: number;
    /** ชื่ออาชีพ (ภาษาอังกฤษ) */
    occupationNameEn: string;
    /** ชื่ออาชีพ (ภาษาไทย) */
    occupationNameTh: string;
    /** custom ประเภททักษะ (ภาษาอังกฤษ) */
    occupationSkills?: MasOccupationSkillData[];
    generalSkillNameTh?:string;
    generalSkillNameEn?:string;
    courseNameTh?:string; 
    courseNameEn?:string;
    jobSelected?:string;
    rowNum?: number;
    size?: number;
}